# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/trips": {
      "get": {
        "operationId": "TripsController_findAll",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role-based access header. Pass one of: Super User, Traveler, Travel Partner, Tour Guide, Vendor, Support Executive.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "Super User",
                "Traveler",
                "Travel Partner",
                "Tour Guide",
                "Vendor",
                "Support Executive"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List all trips."
          }
        },
        "tags": [
          "trips"
        ]
      },
      "post": {
        "operationId": "TripsController_create",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role-based access header. Pass one of: Super User, Traveler, Travel Partner, Tour Guide, Vendor, Support Executive.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "Super User",
                "Traveler",
                "Travel Partner",
                "Tour Guide",
                "Vendor",
                "Support Executive"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTripDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Create a trip directly."
          }
        },
        "tags": [
          "trips"
        ]
      }
    },
    "/trips/{id}": {
      "get": {
        "operationId": "TripsController_findOne",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role-based access header. Pass one of: Super User, Traveler, Travel Partner, Tour Guide, Vendor, Support Executive.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "Super User",
                "Traveler",
                "Travel Partner",
                "Tour Guide",
                "Vendor",
                "Support Executive"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Find trip by id."
          }
        },
        "tags": [
          "trips"
        ]
      },
      "patch": {
        "operationId": "TripsController_update",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role-based access header. Pass one of: Super User, Traveler, Travel Partner, Tour Guide, Vendor, Support Executive.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "Super User",
                "Traveler",
                "Travel Partner",
                "Tour Guide",
                "Vendor",
                "Support Executive"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTripDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Update trip."
          }
        },
        "tags": [
          "trips"
        ]
      },
      "delete": {
        "operationId": "TripsController_remove",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role-based access header. Pass one of: Super User, Traveler, Travel Partner, Tour Guide, Vendor, Support Executive.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "Super User",
                "Traveler",
                "Travel Partner",
                "Tour Guide",
                "Vendor",
                "Support Executive"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Delete trip."
          }
        },
        "tags": [
          "trips"
        ]
      }
    },
    "/trips/{id}/assign-guide": {
      "post": {
        "operationId": "TripsController_assignGuide",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role-based access header. Pass one of: Super User, Traveler, Travel Partner, Tour Guide, Vendor, Support Executive.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "Super User",
                "Traveler",
                "Travel Partner",
                "Tour Guide",
                "Vendor",
                "Support Executive"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignGuideDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "trips"
        ]
      }
    },
    "/trips/{id}/guide/accept": {
      "post": {
        "operationId": "TripsController_acceptGuide",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role-based access header. Pass one of: Super User, Traveler, Travel Partner, Tour Guide, Vendor, Support Executive.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "Super User",
                "Traveler",
                "Travel Partner",
                "Tour Guide",
                "Vendor",
                "Support Executive"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "trips"
        ]
      }
    },
    "/trips/{id}/guide/reject": {
      "post": {
        "operationId": "TripsController_rejectGuide",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role-based access header. Pass one of: Super User, Traveler, Travel Partner, Tour Guide, Vendor, Support Executive.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "Super User",
                "Traveler",
                "Travel Partner",
                "Tour Guide",
                "Vendor",
                "Support Executive"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "trips"
        ]
      }
    },
    "/trips/{id}/assign-vendor": {
      "post": {
        "operationId": "TripsController_assignVendor",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role-based access header. Pass one of: Super User, Traveler, Travel Partner, Tour Guide, Vendor, Support Executive.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "Super User",
                "Traveler",
                "Travel Partner",
                "Tour Guide",
                "Vendor",
                "Support Executive"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignVendorDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "trips"
        ]
      }
    },
    "/trips/{id}/vendor/accept": {
      "post": {
        "operationId": "TripsController_acceptVendor",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role-based access header. Pass one of: Super User, Traveler, Travel Partner, Tour Guide, Vendor, Support Executive.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "Super User",
                "Traveler",
                "Travel Partner",
                "Tour Guide",
                "Vendor",
                "Support Executive"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "trips"
        ]
      }
    },
    "/trips/{id}/vendor/reject": {
      "post": {
        "operationId": "TripsController_rejectVendor",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role-based access header. Pass one of: Super User, Traveler, Travel Partner, Tour Guide, Vendor, Support Executive.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "Super User",
                "Traveler",
                "Travel Partner",
                "Tour Guide",
                "Vendor",
                "Support Executive"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "trips"
        ]
      }
    },
    "/trips/{id}/start": {
      "post": {
        "operationId": "TripsController_start",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role-based access header. Pass one of: Super User, Traveler, Travel Partner, Tour Guide, Vendor, Support Executive.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "Super User",
                "Traveler",
                "Travel Partner",
                "Tour Guide",
                "Vendor",
                "Support Executive"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "trips"
        ]
      }
    },
    "/trips/{id}/schedule": {
      "get": {
        "operationId": "TripsController_schedule",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role-based access header. Pass one of: Super User, Traveler, Travel Partner, Tour Guide, Vendor, Support Executive.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "Super User",
                "Traveler",
                "Travel Partner",
                "Tour Guide",
                "Vendor",
                "Support Executive"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "trips"
        ]
      }
    },
    "/trips/{id}/schedule/{scheduleItemId}": {
      "patch": {
        "operationId": "TripsController_updateSchedule",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role-based access header. Pass one of: Super User, Traveler, Travel Partner, Tour Guide, Vendor, Support Executive.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "Super User",
                "Traveler",
                "Travel Partner",
                "Tour Guide",
                "Vendor",
                "Support Executive"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "scheduleItemId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleProgressDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "trips"
        ]
      }
    },
    "/trips/{id}/guide-updates": {
      "post": {
        "operationId": "TripsController_guideUpdate",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role-based access header. Pass one of: Super User, Traveler, Travel Partner, Tour Guide, Vendor, Support Executive.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "Super User",
                "Traveler",
                "Travel Partner",
                "Tour Guide",
                "Vendor",
                "Support Executive"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GuideUpdateDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "trips"
        ]
      }
    },
    "/trips/{id}/vendor-updates": {
      "post": {
        "operationId": "TripsController_vendorUpdate",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role-based access header. Pass one of: Super User, Traveler, Travel Partner, Tour Guide, Vendor, Support Executive.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "Super User",
                "Traveler",
                "Travel Partner",
                "Tour Guide",
                "Vendor",
                "Support Executive"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VendorUpdateDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "trips"
        ]
      }
    },
    "/trips/{id}/updates": {
      "get": {
        "operationId": "TripsController_updates",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role-based access header. Pass one of: Super User, Traveler, Travel Partner, Tour Guide, Vendor, Support Executive.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "Super User",
                "Traveler",
                "Travel Partner",
                "Tour Guide",
                "Vendor",
                "Support Executive"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "trips"
        ]
      }
    },
...
  },
  "info": {
    "title": "Dream Destination Review 4 API",
    "description": "NestJS in-memory REST API with RBAC header authorization for traveler, travel partner, tour guide, vendor, support, and super user portals.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "x-role": {
        "type": "apiKey",
        "in": "header",
        "name": "x-role",
        "description": "Role header required on every API. Use Super User, Traveler, Travel Partner, Tour Guide, Vendor, or Support Executive."
      }
    },
    "schemas": {
...
      "CreateTripDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Goa Beach Tour"
          },
          "destination": {
            "type": "string",
            "example": "Goa, India"
          },
          "travelerName": {
            "type": "string",
            "example": "Rahul Sharma"
          },
          "travelerEmail": {
            "type": "string",
            "example": "rahul@example.com"
          },
          "startDate": {
            "type": "string",
            "example": "2026-05-12"
          },
          "endDate": {
            "type": "string",
            "example": "2026-05-16"
          },
          "adults": {
            "type": "number",
            "example": 2
          },
          "children": {
            "type": "number",
            "example": 0
          },
          "budget": {
            "type": "number",
            "example": 4500
          },
          "status": {
            "type": "string",
            "example": "planning"
          },
          "interests": {
            "example": [
              "beach",
              "heritage"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "notes": {
            "type": "string",
            "example": "Need family-friendly activities."
          }
        },
        "required": [
          "title",
          "destination",
          "travelerName"
        ]
      },
      "UpdateTripDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Goa Beach Tour"
          },
          "destination": {
            "type": "string",
            "example": "Goa, India"
          },
          "travelerName": {
            "type": "string",
            "example": "Rahul Sharma"
          },
          "travelerEmail": {
            "type": "string",
            "example": "rahul@example.com"
          },
          "startDate": {
            "type": "string",
            "example": "2026-05-12"
          },
          "endDate": {
            "type": "string",
            "example": "2026-05-16"
          },
          "adults": {
            "type": "number",
            "example": 2
          },
          "children": {
            "type": "number",
            "example": 0
          },
          "budget": {
            "type": "number",
            "example": 4500
          },
          "status": {
            "type": "string",
            "example": "planning"
          },
          "interests": {
            "example": [
              "beach",
              "heritage"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "notes": {
            "type": "string",
            "example": "Need family-friendly activities."
          }
        }
      },
      "AssignGuideDto": {
        "type": "object",
        "properties": {
          "guideId": {
            "type": "string",
            "example": "GUIDE-1"
          },
          "guideName": {
            "type": "string",
            "example": "Arun Kumar"
          }
        }
      },
      "AssignVendorDto": {
        "type": "object",
        "properties": {
          "vendorId": {
            "type": "string",
            "example": "VENDOR-1"
          },
          "vendorName": {
            "type": "string",
            "example": "ABC Travels"
          },
          "serviceType": {
            "type": "string",
            "example": "Transport"
          }
        }
      },
      "ScheduleProgressDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "completed",
            "enum": [
              "in-progress",
              "completed",
              "delay"
            ]
          },
          "notes": {
            "type": "string",
            "example": "Pickup completed successfully."
          },
          "location": {
            "type": "string",
            "example": "Goa Airport"
          },
          "updatedBy": {
            "type": "string",
            "example": "ABC Travels"
          }
        },
        "required": [
          "status"
        ]
      },
      "GuideUpdateDto": {
        "type": "object",
        "properties": {
          "scheduleItemId": {
            "type": "string",
            "example": "SCH-TRIP-101-2"
          },
          "status": {
            "type": "string",
            "example": "completed",
            "enum": [
              "location",
              "started",
              "delay",
              "completed"
            ]
          },
          "statusText": {
            "type": "string",
            "example": "Tour Completed"
          },
          "location": {
            "type": "string",
            "example": "Calangute"
          },
          "notes": {
            "type": "string",
            "example": "Traveler reached the current activity."
          }
        },
        "required": [
          "scheduleItemId",
          "status"
        ]
      },
      "VendorUpdateDto": {
        "type": "object",
        "properties": {
          "scheduleItemId": {
            "type": "string",
            "example": "SCH-TRIP-101-1"
          },
          "status": {
            "type": "string",
            "example": "completed",
            "enum": [
              "progress",
              "enroute",
              "completed"
            ]
          },
          "statusText": {
            "type": "string",
            "example": "Completed"
          },
          "serviceType": {
            "type": "string",
            "example": "Transport"
          },
          "message": {
            "type": "string",
            "example": "Transport pickup completed."
          },
          "location": {
            "type": "string",
            "example": "Goa Airport"
          }
        },
        "required": [
          "scheduleItemId",
          "status",
          "message"
        ]
      },
...
    }
  }
}